fix client-only classes usage on server
